ДЗ. Модуляция.
